Update Ownership.h
Added two new enumerated type elements to allow for case where an action deletes something without checking it as a precondition and a second action preconditions and deletes the same proposition. This should lead to a mutex violation, but depending on order of use in a plan, did not reliably do so.